Added the batch file start-compass-watch.bat
start-compass-watch.bat
=======================
The purpose of this script is to call "compass watch"
"compass watch" Compiles Sass stylesheets to CSS when they change
If the config.rb file and this script are in the same folder the
script will call compass watch. If they are not in the same folder,
the script will look 1 directory above it's folder for the config.rb
file.
echo